pyproject.toml
[tool.poetry]
name = "gsb-selenium"
version = "1.0.0"
description = "Google Search Bot using Selenium - Advanced browser automation for search marketing"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "gsb_selenium"}]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.15.0"
selenium-wire = "^5.1.0"
webdriver-manager = "^4.0.1"
selenium-stealth = "^1.0.6"
python-dotenv = "^1.0.0"
aiofiles = "^23.2.1"
aiohttp = "^3.9.5"
loguru = "^0.7.2"
asyncclick = "^8.1.7.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
2captcha-python = "^1.4.0"
pymongo = "^4.6.0"
boto3 = "^1.34.0"
psutil = "^7.0.0"
undetected-chromedriver = "^3.5.4"
fake-useragent = "^1.4.0"
pyautogui = "^0.9.54"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gsb = "gsb_selenium.cli:main"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true